packages/pynumaflow/pyproject.toml
[project]
name = "pynumaflow"
version = "0.12.1"
description = "Provides the interfaces of writing Python User Defined Functions and Sinks for NumaFlow."
authors = [
    { name = "NumaFlow Developers" },
]
maintainers = [
    { name = "Avik Basu", email = "avikbasu93@gmail.com" },
    { name = "Vigith Maurice", email = "vigith@gmail.com" },
]
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10,<3.15"
classifiers = [
    "Topic :: Software Development :: Libraries",
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "grpcio>=1.48.1",
    "grpcio-tools>=1.48.1",
    "google-api-core>=2.11.0",
    "grpcio-status>=1.48.1",
    "protobuf>=6.31.1,<7.0",
    "aiorun>=2025.1.1",
    "uvloop>=0.21.0",
    "psutil>=6.0.0",
]

[project.urls]
Repository = "https://github.com/numaproj/numaflow-python"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=3.0",
    "black>=23.1",
    "grpcio-testing>=1.48.1",
    "ruff>=0.0.264",
    "pre-commit>=3.3.1",
    "types-protobuf>=6.32.1.20250918",
    "mypy>=1.18.2",
    "grpc-stubs>=1.53.0.6",
    "types-psutil>=7.0.0.20251001",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material[imaging]>=9.5",
    "mkdocs-exclude>=1.0",
    "mkdocstrings[python]>=0.27",
    "mike>=2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | \.idea
  | _build
  | buck-out
  | build
  | dist
  | tests/.*/setup.py
  | function/proto
  | sink/proto
)/
'''

[tool.pytest]
strict = true
collect_imported_tests = false
console_output_style = "times"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
markers = [
    "integration: tests that start real gRPC servers on Unix sockets (slower)",
]

[tool.ruff]
line-length = 100
src = ["pynumaflow", "tests", "examples"]
target-version = "py310"
show-fixes = true
extend-exclude = [
    "*_pb2.py",
    "*_pb2*.py",
    "*.pyi"
]

[tool.ruff.lint]
select = ["E", "F", "UP"]
